Add files via upload
For the proper functioning of this code, the second column must contain the values that describe the empty values in the other columns. This is useful because Excel and DAX do not allow for this calculation in an easy way. With this approach, we can dynamically relate the columns to the necessary values.